adding UML in img

+++ tenmo-server/src/main/java/com/techelevator/tenmo/model/Account.java
@@ -5,33 +5,33 @@ import com.techelevator.tenmo.exception.InsufficientBalanceException;
 import java.math.BigDecimal;
 
 public class Account {
-    private long account_id;
-    private long user_id;
+    private int account_id;
+    private int user_id;
     private BigDecimal balance;
 
     public Account() {
 
     }
 
-    public Account(long account_id, long user_id, BigDecimal balance) {
+    public Account(int account_id, int user_id, BigDecimal balance) {
         this.account_id = account_id;
         this.user_id = user_id;
         this.balance = balance;
     }
 
-    public long getAccount_id() {
+    public int getAccount_id() {
         return account_id;
     }
 
-    public void setAccount_id(long account_id) {
+    public void setAccount_id(int account_id) {
         this.account_id = account_id;
     }
 
-    public long getUser_id() {
+    public int getUser_id() {
         return user_id;
     }
 
-    public void setUser_id(long user_id) {
+    public void setUser_id(int user_id) {
         this.user_id = user_id;
     }
